refactor: Button -> PrimaryButton, SmallButton -> SecondaryButton 으로 네이밍 수정 및 width props 추가
- 버튼의 재사용성과 공통 컴포넌트로써의 역할을 명확히 하기 위해 네이밍 수정
