+++ src/components/@common/Button/PrimaryButton/index.tsx
@@ -1,4 +1,4 @@
-import * as Styled from './Button.styled';
+import * as Styled from './PrimaryButton.styled';
 
 export interface IButtonProps
   extends React.ButtonHTMLAttributes<HTMLButtonElement> {}
@@ -13,8 +13,8 @@ export interface IButtonProps
  *      1) 클릭 이벤트 처리: 버튼 클릭 시 특정 작업(이벤트 핸들러 실행)을 수행
  */
 
-const Button = ({ children, ...rest }: IButtonProps) => {
+const PrimaryButton = ({ children, ...rest }: IButtonProps) => {
   return <Styled.Button {...rest}>{children}</Styled.Button>;
 };
 
-export default Button;
+export default PrimaryButton;
