+++ src/components/auth/Form/ProfileForm/index.tsx
@@ -1,4 +1,4 @@
-import Button from '@/components/@common/Button/Button';
+import PrimaryButton from '@/components/@common/Button/PrimaryButton';
 import signUpStepAtom from '@/recoil/auth/signUp/atom';
 import { moveToStep } from '@/utils/step/moveSteps';
 import { useSetRecoilState } from 'recoil';
@@ -7,9 +7,12 @@ const ProfileForm = () => {
   const setSignUpStep = useSetRecoilState(signUpStepAtom);
   return (
     <>
-      <Button type="button" onClick={() => moveToStep('prev', setSignUpStep)}>
+      <PrimaryButton
+        type="button"
+        onClick={() => moveToStep('prev', setSignUpStep)}
+      >
         뒤로
-      </Button>
+      </PrimaryButton>
     </>
   );
 };
